:recycle: [functional] Move `move_repository_files_to_subdirectory` to tests.util.git
