+++ tests/functional/conftest.py
@@ -69,13 +69,3 @@ def repository(template_directory: Path) -> Path:
     repository = pygit2.init_repository(template_directory)
     commit(repository, message="Initial")
     return template_directory
-
-
-def move_repository_files_to_subdirectory(repositorypath: Path, directory: str) -> None:
-    """Move all files in the repository to the given subdirectory."""
-    repository = pygit2.Repository(repositorypath)
-    builder = repository.TreeBuilder()
-    builder.insert(directory, repository.head.peel().tree.id, pygit2.GIT_FILEMODE_TREE)
-    tree = repository[builder.write()]
-    repository.checkout_tree(tree)
-    commit(repository, message=f"Move files to subdirectory {directory}")
